+++ dflib/src/test/java/com/nhl/dflib/DataFrame_AddDropMapColumnsTest.java
@@ -133,6 +133,19 @@ public class DataFrame_AddDropMapColumnsTest {
     }
 
     @Test
+    public void testDropColumns_Predicate() {
+        DataFrame df = DataFrame.newFrame("a1", "b2", "c1")
+                .foldByColumn(1, 2, 3, 4, 5, 6, 7, 8, 9)
+                .dropColumns(c -> c.endsWith("1"));
+
+        new DataFrameAsserts(df, "b2")
+                .expectHeight(3)
+                .expectRow(0, 4)
+                .expectRow(1, 5)
+                .expectRow(2, 6);
+    }
+
+    @Test
     public void testAddRowNumber() {
         DataFrame df = DataFrame.newFrame("a", "b").foldByRow(
                 1, "x",